Integrate with Algolia DocSearch. (#23)
> Wow, this looks amazing! Thank you so much for setting this up. (Really sorry about my delayed reply - the past week has been a bit hectic for me...)
> 
> I can't wait to try the incredible new search box :-)
> 

No worries at all, thanks for taking a look! I also had pretty packed weeks earlier, sorry to leave the docs migration and releasing work there for so long, now I have more time after hours, am planning to finish this piece of work as soon as by Nov.

> > Maybe we should start adding policy/privacy banners such as below to the site about the cookie usage.
> 
> That sounds reasonable! I guess we don't have to worry too much about that until one day we get a lot of traffic :-)

Cool, that's also I was thinking too, will create an issue to document the future work but will leave it as-is for now.